# docs/swagger.yaml
basePath: /api/v1
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  model.IDC:
    properties:
      code:
        type: string
      contact:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      locate:
        type: string
      name:
        type: string
      orderSite:
        type: string
      remark:
        type: string
      telephone:
        type: string
      updatedAt:
        type: string
      vendor:
        description: 供应商信息
        type: string
      website:
        type: string
    type: object
  model.PaginationResult-array_model_IDC:
    properties:
      count:
        type: integer
      data:
        items:
          $ref: '#/definitions/model.IDC'
        type: array
    type: object
  restful.ResponseBody-model_IDC:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/model.IDC'
      msg:
        type: string
    type: object
  restful.ResponseBody-model_PaginationResult-array_model_IDC:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/model.PaginationResult-array_model_IDC'
      msg:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /idc:
    get:
      consumes:
      - application/json
      description: Get IDC List
      parameters:
      - description: IDC Code
        in: query
        name: code
        type: string
      - description: IDC Name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/restful.ResponseBody-model_PaginationResult-array_model_IDC'
      summary: Get IDC List
      tags:
      - idc
    post:
      consumes:
      - application/json
      description: Create IDC
      parameters:
      - description: IDC
        in: body
        name: idc
        required: true
        schema:
          $ref: '#/definitions/model.IDC'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/restful.ResponseBody-model_IDC'
      summary: Create IDC
      tags:
      - idc
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"